mgr/dashboard,prometheus: fix handling of server_addr
Commit ed1bc1c corrected the URI published by dashboard/prometheus
to use get_mgr_ip() if the address was the wildcard (eg ::).
However, the change also affected the cherrypy server.socket_host
config value, so the modules could no longer bind to a wildcard.

This patch corrects the commit to not affect the cherrypy config.

It also further corrects the dashboard URI fix to handle the 0.0.0.0
wildcard case (prometheus already had this behavior).

Fixes: ed1bc1c0ab1f92862bc3aa019d17d7e84610510c
Fixes: http://tracker.ceph.com/issues/52002
Signed-off-by: Scott Shambarger <devel@shambarger.net>