osd/scrub: remove the fast track (w/o resources acquisition) for repair scrubs
Previously, after-repair scrubs were started without waiting for either local
or remote OSDs' scrub resources.
The tagging of scrub sessions by the replicas is based on monitoring replica-request
and replica-release messages from the primary. Scrub-map requests arriving without
any reservations interfere with this mechanism.
The benefits of this fast-track were limited at best, and do not justify the complexity
of a solution that accommodates both the bypass and the tagging.

Signed-off-by: Ronen Friedman <rfriedma@redhat.com>