mgr/dashboard: Fix failing config dashboard e2e check
Recently a new osd config has been added in
dcd68a812eba6a7aa3c4d7a2edd4751be5388868 and this is getting Modified.
So on our dashboard e2e config check which checks for the Modified
filter, this is also coming on the entry. So we need to increase the
count.

Fixes: https://tracker.ceph.com/issues/52649
Signed-off-by: Nizamudeen A <nia@redhat.com>