seastore : implement multi-stream write of NVMeDevice class
For checking an SSD supports multi-stream or not, add identifing interfaces.
identify_controller/namespace command provides feature information which
SSD provides. If the feature is not supported, write() ignores stream ID.
Because there is no interface to know that the range of valid stream ID is not
provided by NVMe specification. For the stream ID which is out of supported
range, SSD regards it as on-streamd IO.

Signed-off-by: Jinyong Ha <jy200.ha@samsung.com>