msg/async: support disabling data crc for protocol v2
In protocol v1, the ms_crc_data config option allowed for globally disabling
the crc calculation and verification for data segments.
This option was so far ignored in protocol v2 implementation.
This commit extends v2 to respect this configuration option.
When the data crc is disabled, zeros will be sent instead of the supposed calculated crc.

Signed-off-by: Or Ozeri <oro@il.ibm.com>