doc/cephadm: add admonition for tracker 53062
NOTE: This commit also adds an admonition to the pacific.rst
      page, in the /releases directory.

This commit warns users not to upgrade to Ceph v16 (Pacific). Upgrading
to v16 (Pacific) can cause data corruption.

This commit has been made in response to https://tracker.ceph.com/issues/53062
and in response to Stefan Kooman's urgent request that we update the
documentation, adding a warning to the documentation to prevent users
from upgrading to v16.2.6.

Signed-off-by: Zac Dover <zac.dover@gmail.com>