ceph-volume: adjust arguments for 'ceph-volume raw activate'
Take a list of devices, so that we can selectively activate a raw osd
with db/wal.

Remove the argument type kludge introduced in c75f09d626efc30428532509c2d77011807a80e2
since it is no longer needed.

Note that we're making this change because (1) it allows db/wal and (2)
because there are no known users of 'raw activate'.  The only known user
is via 'ceph-volume activate' and we've fixed that caller in this commit.

Signed-off-by: Sage Weil <sage@newdream.net>