doc/governance: update based on review and CLT discussions
Clarified some parts of council that were discussed previously,
specifying the number of members and a staggered term.

Added a bit more about the steering committee - thinking the meetings
could be split between tactical (3/4 weeks) and strategic (monthly),
and still open to anyone to join the discussion, but restricted to
only members voting.

Removed the meeting section since that belongs more in a separate
place, like the ceph.io website.

Signed-off-by: Josh Durgin <jdurgin@redhat.com>