doc/dpdk: reword the root access part
the root access to system is not a must have for running a DPDK
application. so reword the "Configuring OSD DPDKStack" section.

also, manually editing /etc/passwd is not encouraged, so use
"usermod" instead. to add a directory after user's command
interpreter in /etc/passwd does not make sense. see PASSWD(5).
so drop the paragraph on editing /etc/passwd.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>