rgw: add missing virtual decls for dtors
As a side effect or benefit of zipper loadable module work have
found several classes that are missing virtual decl on dtors, or
are missing a dtor decl entirely.

In the zipper loadable module work these missing decls result in
undefined references the vtable for the various classes during the
link.

Signed-off-by: Kaleb S. KEITHLEY <kkeithle@redhat.com>