crimson/osd: add a type wrapper for PG and OSD-hosted OSDMapGate instances.
`ClientRequest` is going to hold all tracking events as a `std::tuple`
instance and access them through the type-related variant of `std::get`.
As two instances of `OSDMapGate` are used there, a need to differentiate
at the type level arises. This commit addresses that need without
templatazing the already existing class and code shuffling.

Signed-off-by: Radosław Zarzyński <rzarzyns@redhat.com>