doc/rados/api: fix invalid RST syntax and don't sudo for vim
Fix invalid syntax where "linenos:" was printed in the
final rendered documentation instead of being used as
formatting syntax.

There is no need to use sudo for editing a source file
so run vim without sudo.

Change the whole block to use bash prompt since it is
a command.

Also modify the preceding text to hopefully not bold it.

Signed-off-by: Ville Ojamo <14869000+bluikko@users.noreply.github.com>