crimson/osd: reset logger before exit
* extract the code to set logging fstream into a dedicated function
* do not reset logging until the end of the seastar application.

before this change, `reset_logger` is created in the
`if (auto log_file = local_conf()->log_file; !log_file.empty())` branch,
so its life cycle ends when the `if` block ends. in other words,
the cerr fstream is used for logging after the `if` block ends.
this is not the expected behavior.

after this changge, `reset_logger` is created out of the `if` block.
so we won't reset the logger back to `cerr` until the lambda passed to
`seastar::async()` exits.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>