mgr/balancer: Disallow 7, 2400 as weekday/time
Since PR 37941 was merged, the OSD Scrub Schedule no longer accepts 7
and 2400 as values for the weekday and time respectively, but the
balancer module in the manager still uses the old parameters. For
consistency, this commit swaps the balancer to also reject days with the
value of 7.

This change also modifies the behaviour of the
{begin,end}_{weekday,time} settings to match that in the OSD Scrub
Scheduler - As 7 is no longer permitted, setting the two values to be
equal will set the window to be always on.

Fixes: https://tracker.ceph.com/issues/49086
Signed-off-by: Tom Coldrick <tcoldrick@bloomberg.net>