mds: Don't blocklist clients in any replay state
The mds in 'standby-replay' state should not blocklist clients.
Attempting to do so would hit an assert while journaling.

Before the commit a74ff15 the 'Server:apply_blocklist' used to
fetch the blocklist addresses from 'Objecter::consume_blocklist_events'
which returns blocklist addresses only if 'blocklist_events_enabled'
are enabled. This is not enabled for mds in 'standby-replay' state.
Hence this issue wasn't seen.

After the commit a74ff15 the 'Server::apply_blocklist' fetches the
blocklist addresses from the osdmap. This would return the blocklist
addresses irrespective of the mds state. The mds in 'replay' state
doesn't blocklist the clients but if the mds is in 'standby-replay'
state, it tries to journal the blocklist and asserts. This patch
fixes the same by not allowing the blocklist if the mds is in any
replay state.

Introduced-by: a74ff153afc49674dec4de8df723bbda7c2825a5
Fixes: https://tracker.ceph.com/issues/56012
Signed-off-by: Kotresh HR <khiremat@redhat.com>