librbd/crypto: fix read return check
handle_read so far assumed that a successful read returns 0.
This is true for the core layer which returns 0 at ObjectRequest.cc.
However, this is not the case for ParentCacheObjectDispatch,
which returns the (possibly positive) size of data read, instead of 0.
To account for that, we change the success check from r==0 to r>=0.

Signed-off-by: Or Ozeri <oro@il.ibm.com>