client: abort the client if we couldn't invalidate dentry caches
The option 'client_die_on_failed_dentry_invalidate' requires to kill
the client when fails to invalidate the dentry caches from kernel.
The CephFS client requires a mechanism to invalidate dentries in the
caller (e.g. the kernel for ceph-fuse) when capabilities must be recalled.
If the client cannot do this then the MDS cache cannot shrink which
can cause the MDS to fail.

Introduced by commit 9252cfe8e00cd97ae36bce273048470768bcb673.

Fixes: https://tracker.ceph.com/issues/57126
Signed-off-by: Xiubo Li <xiubli@redhat.com>