doc/rados: update prompts in crush-map-edits.rst
This PR updates the prompts in crush-map-edits.rst
to make them unselectable.

There remains no good known way to render parts of
a file if the first line of that file begins with a
hash (#). Sphinx italicizes such a first line, which
is not what I want. Two examples of this are present
in the file crush-maps-rst under the section called
"CRUSH Map Bucket Types". I set this down here for
my own records, in case it is helpful in finding a
way to make these sections render as I would prefer.

Signed-off-by: Zac Dover <zac.dover@gmail.com>